+++ packages/flopper/index.js
@@ -1,4 +1,7 @@
-export { palettFlopper }                                               from './src/palettFlopper.js'
-export { palettFlopperLite }                                           from './src/palettFlopperLite.js'
-export { presetFlopperMaterial }                                       from './src/presetFlopperMaterial.js'
-export { presetFlopperPavtone, presetFlopperPavtone as presetFlopper } from './src/presetFlopperPavtone.js'
+export { palettFlopper }         from './src/palettFlopper.js'
+export { palettFlopperLite }     from './src/palettFlopperLite.js'
+export { presetFlopperMaterial } from './src/presetFlopperMaterial.js'
+
+export {
+  presFlopper, presFlopper as presetFlopperPavtone, presFlopper as presetFlopper
+} from './src/presFlopper.js'