+++ packages/flopper/src/palettFlopper.js
@@ -3,9 +3,9 @@ import { NUM_DESC }                            from '@aryth/comparer'
 import { Grey }                                from '@palett/cards'
 import { HEX }                                 from '@palett/enum-color-space'
 import { ColorGroups, Degrees, palettCrostab } from '@palett/table'
-import { swap }                                from '@vect/swap'
-import { degreeToIndice }                      from '../utils/degreeToIndice.js'
-import { sortBy }                              from '../utils/sortDegrees.js'
+import { swap }           from '@vect/swap'
+import { degreeToIndice } from './utils/degreeToIndice.js'
+import { sortBy }         from './utils/sortDegrees.js'
 
 export function* palettFlopper({
                                  degrees = Degrees.entire,