+++ packages/flopper/CHANGELOG.md
@@ -1,5 +1,47 @@
 # @palett/flopper
 
+## 0.12.9
+
+### Patch Changes
+
+- fix bugs
+- Updated dependencies
+  - @palett/enum-color-space@0.12.9
+  - @palett/color-space@0.12.9
+  - @palett/convert@0.12.9
+  - @palett/munsell@0.12.9
+  - @palett/presets@0.12.9
+  - @palett/cards@0.12.9
+  - @palett/table@0.12.9
+
+## 0.12.8
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @palett/enum-color-space@0.12.8
+  - @palett/color-space@0.12.8
+  - @palett/convert@0.12.8
+  - @palett/munsell@0.12.8
+  - @palett/presets@0.12.8
+  - @palett/cards@0.12.8
+  - @palett/table@0.12.8
+
+## 0.12.7
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @palett/enum-color-space@0.12.7
+  - @palett/cards@0.12.7
+  - @palett/color-space@0.12.7
+  - @palett/convert@0.12.7
+  - @palett/munsell@0.12.7
+  - @palett/presets@0.12.7
+  - @palett/table@0.12.7
+
 ## 0.12.6
 
 ### Patch Changes